[web] add templatenum limit #188

+++ web/src/views/pages/features/template/ModelList.vue
@@ -1,5 +1,5 @@
 <script>
-import { ref, computed, onMounted } from 'vue';
+import { ref, computed, onMounted, watch } from 'vue';
 import Annotation from '@/components/Annotation.vue';
 import moment from 'moment';
 import { ElMessageBox, ElMessage } from 'element-plus';
@@ -8,6 +8,7 @@ import useAnnotator from '@/mixins/useAnnotator.js';
 import { useStore } from 'vuex';
 import { useRouter } from 'vue-router';
 import { apiClient } from '@/service/auth.js';
+import { templateLimit } from '@/constants.js';
 import { handleErrorMsg } from '@/mixins/useCommon.js';
 import { Delete, Edit, Download, Pointer, View } from '@element-plus/icons-vue';
 import EditableRow from '@/components/EditableRow.vue';
@@ -46,8 +47,10 @@ export default {
 
         const images = ref(null);
         const tableData = ref([]);
+        const templateNum = ref(0);
         const dialogVisible = ref(false);
         const dialogWidth = ref('');
+        const blockCreate = ref(false);
 
         const breadcrumbItems = ref([{ path: '/', label: '首頁' }, { label: '通用辨識' }, { label: '模板辨識', isCurrent: true }]);
         const actions = ref([
@@ -57,25 +60,6 @@ export default {
             { label: '刪除', icon: 'Delete', type: 'danger', handler: 'handleDelete' },
             { label: '下載', icon: 'Download', type: 'info', handler: 'downloadTemplate' }
         ]);
-
-        onMounted(async () => {
-            images.value = await galleriaService.getImages();
-            tableData.value = await getAvailableTemplate();
-        });
-
-        const formattedTableData = computed(() => {
-            if (tableData.value.length === 0) {
-                return [];
-            }
-            return tableData.value.map((row, idx) => {
-                return {
-                    idx: idx + 1,
-                    ...row,
-                    created_at: formatDate(row.creation_time),
-                    expired_at: formatDate(row.expiration_time)
-                };
-            });
-        });
 
         const selectedRectangleType = ref({
             name: '文字',
@@ -115,6 +99,30 @@ export default {
         const creation_time = ref('');
         const expiration_time = ref('');
         const buttonText = ref('新增模板');
+
+        onMounted(async () => {
+            images.value = await galleriaService.getImages();
+            tableData.value = await getAvailableTemplate();
+        });
+
+        const formattedTableData = computed(() => {
+            if (tableData.value.length === 0) {
+                return [];
+            }
+            return tableData.value.map((row, idx) => {
+                return {
+                    idx: idx + 1,
+                    ...row,
+                    created_at: formatDate(row.creation_time),
+                    expired_at: formatDate(row.expiration_time)
+                };
+            });
+        });
+
+        watch(formattedTableData, (newValue, oldValue) => {
+            templateNum.value = newValue.length;
+            blockCreate.value = templateNum.value >= templateLimit ? true : false;
+        });
 
         // Methods
         function formatDate(date) {
@@ -220,6 +228,13 @@ export default {
         }
 
         function createTemplate() {
+            if (blockCreate.value) {
+                ElMessage({
+                    type: 'error',
+                    message: '已達模板數量上限，請刪除模板後再重新新增。'
+                });
+                return;
+            }
             sessionStorage.clear();
             store.commit('createNewUpdate', true);
             store.commit('templateNameUpdate', '');
@@ -343,6 +358,9 @@ export default {
             initialData,
             creation_time,
             expiration_time,
+            numLimit: templateLimit,
+            templateNum,
+            blockCreate,
             formatDate,
             getAvailableTemplate,
             handleConfirm,
@@ -368,34 +386,30 @@ export default {
 </script>
 <template>
     <div class="layoutZoneContainer">
-
         <div class="action-header">
-            <p class="title" style="margin-bottom: 0px">模板列表</p>
+            <div class="title-block">
+                <p class="title" style="margin-bottom: 0px; margin-right: 0.5rem;">模板列表</p>
+                <p class="sub-info">模板個數：</p>
+                <p class="sub-info" :class="{ 'red-text': blockCreate }">{{ templateNum }}</p>
+                <p class="sub-info">/{{ numLimit }}</p>
+            </div>
             <button class="uiStyle sizeS subLength btnGreen" @click="createTemplate">
                 {{ buttonText }}
             </button>
         </div>
 
-        <el-table :data="formattedTableData" style="width: 100%" border :default-sort = "{prop: 'template_id', order: 'descending'}">
-            <el-table-column v-for="item in tableHeader" :key="item.prop" :prop="item.prop" :label="item.label" :min-width="item.width">
+        <el-table :data="formattedTableData" style="width: 100%" border
+            :default-sort="{ prop: 'template_id', order: 'descending' }">
+            <el-table-column v-for="item in tableHeader" :key="item.prop" :prop="item.prop" :label="item.label"
+                :min-width="item.width">
                 <template #default="scope">
                     <EditableRow :item="item" :row="scope.row" @edit="handleEdit(scope.$index, scope.row)" />
                 </template>
             </el-table-column>
 
-            <ActionColumn
-                v-for="action in actions"
-                :key="action.label"
-                :label="action.label"
-                :icon="action.icon"
-                :type="action.type"
-                :handler="action.handler"
-                @templateOCR="templateOCR"
-                @handleLook="handleLook"
-                @editTemplate="editTemplate"
-                @handleDelete="handleDelete"
-                @downloadTemplate="downloadTemplate"
-            />
+            <ActionColumn v-for="action in actions" :key="action.label" :label="action.label" :icon="action.icon"
+                :type="action.type" :handler="action.handler" @templateOCR="templateOCR" @handleLook="handleLook"
+                @editTemplate="editTemplate" @handleDelete="handleDelete" @downloadTemplate="downloadTemplate" />
         </el-table>
 
         <el-dialog v-model="dialogVisible" :width="dialogWidth">
@@ -406,20 +420,13 @@ export default {
                 <div class="flex flex-column">
                     <div class="flex align-items-center justify-content-center h-4rem font-bold border-round m-2">
                         <!-- Wrap the SelectButton in a container -->
-                        <SelectButton v-model="selectedRectangleType" :options="rectangleTypes" optionLabel="name" @change="handleLook(template_id, selectedRectangleType?.code)" class="selectButton"/>
+                        <SelectButton v-model="selectedRectangleType" :options="rectangleTypes" optionLabel="name"
+                            @change="handleLook(template_id, selectedRectangleType?.code)" class="selectButton" />
                     </div>
                     <div class="flex align-items-center justify-content-center font-bold border-round m-2">
-                        <Annotation
-                            ref="child"
-                            containerId="my-pic-annotation-output"
-                            :editMode="false"
-                            :imageSrc="imageSrc"
-                            width="1000px"
-                            :height="height"
-                            :initialData="initialData"
-                            :justShow="true"
-                            :isVertical="true"
-                        ></Annotation>
+                        <Annotation ref="child" containerId="my-pic-annotation-output" :editMode="false"
+                            :imageSrc="imageSrc" width="1000px" :height="height" :initialData="initialData" :justShow="true"
+                            :isVertical="true"></Annotation>
                     </div>
                     <div class="flex align-items-center justify-content-center font-bold border-round m-2"></div>
                 </div>
@@ -429,12 +436,28 @@ export default {
 </template>
 
 <style scoped>
+.title-block {
+    display: flex;
+    align-items: end;
+}
+
+.sub-info {
+    color: #6c757d;
+    margin: 0;
+}
+
 .select-button-container {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-    align-items: center; /* This will center-align the buttons horizontally */
-    height: 100%; /* Adjust this if you want a specific height */
+    align-items: center;
+    /* This will center-align the buttons horizontally */
+    height: 100%;
+    /* Adjust this if you want a specific height */
+}
+.red-text {
+    color: #e24c4c;
+    font-weight: bold;
 }
 .action-header {
     display: flex;
@@ -442,6 +465,7 @@ export default {
     justify-content: space-between;
     margin-bottom: 20px;
 }
+
 .selectButton ::v-deep .p-button {
     background: #ffffff;
     padding: 4px 3px;
@@ -450,46 +474,55 @@ export default {
     transition: background-color 0.2s, color 0.2s, border-color 0.2s, box-shadow 0.2s;
 }
 
-.selectButton ::v-deep .p-button .p-button-icon-right, .p-button .p-button-icon-left{
+.selectButton ::v-deep .p-button .p-button-icon-right,
+.p-button .p-button-icon-left {
     color: #6c757d;
 }
 
-.selectButton ::v-deep .p-button:not(.p-disabled):not(.p-highlight):hover{
+.selectButton ::v-deep .p-button:not(.p-disabled):not(.p-highlight):hover {
     background: #e9ecef;
     border-color: #ced4da;
     color: #495057;
 }
+
 .selectButton ::v-deep .p-button:not(.p-disabled):not(.p-highlight):hover .p-button-icon-left,
 .selectButton ::v-deep .p-button:not(.p-disabled):not(.p-highlight):hover .p-button-icon-right {
-  color: #343a40;
-}
-.selectButton ::v-deep .p-button.p-highlight {
-  background: #09747A;
-  border-color: #09747A;
-  color: #ffffff;
-}
-.selectButton ::v-deep .p-button.p-highlight .p-button-icon-left,
-.selectButton ::v-deep .p-button.p-highlight .p-button-icon-right {
-  color: #ffffff;
-}
-.selectButton ::v-deep .p-button.p-highlight:hover {
-  background: #10A0A7;
-  border-color: #10A0A7;
-  color: #ffffff;
-}
-.selectButton ::v-deep .p-button.p-highlight:hover .p-button-icon-left,
-.selectButton ::v-deep .p-button.p-highlight:hover .p-button-icon-right {
-  color: #ffffff;
+    color: #343a40;
 }
 
-p-selectbutton.ng-dirty.ng-invalid > .selectButton ::v-deep > .p-button {
-  border-color: #e24c4c;
+.selectButton ::v-deep .p-button.p-highlight {
+    background: #09747A;
+    border-color: #09747A;
+    color: #ffffff;
 }
+
+.selectButton ::v-deep .p-button.p-highlight .p-button-icon-left,
+.selectButton ::v-deep .p-button.p-highlight .p-button-icon-right {
+    color: #ffffff;
+}
+
+.selectButton ::v-deep .p-button.p-highlight:hover {
+    background: #10A0A7;
+    border-color: #10A0A7;
+    color: #ffffff;
+}
+
+.selectButton ::v-deep .p-button.p-highlight:hover .p-button-icon-left,
+.selectButton ::v-deep .p-button.p-highlight:hover .p-button-icon-right {
+    color: #ffffff;
+}
+
+p-selectbutton.ng-dirty.ng-invalid>.selectButton ::v-deep>.p-button {
+    border-color: #e24c4c;
+}
+
 .p-buttonset {
     display: flex;
     justify-content: center;
-    width: 60%; /* adjust this if needed */
-    margin: 0 auto; /* makes the button group center-aligned */
+    width: 60%;
+    /* adjust this if needed */
+    margin: 0 auto;
+    /* makes the button group center-aligned */
 }
 
 .p-highlight {
